Add files via upload
Currently looking at doing a projectile motion simulator. Havent put any of the code in, working it out on paper so far. Looking at whether I should have one object displayed that i'll change the position of or if I should have several objects that pop up at certain intervals. Weird part is that projectile will be in air for variable amounts of time so the intervals displayed would have to vary. I think this means I need to calculate time in air first which requires that I create a checker for when y-coor=0 but doesnt go off right away. May start shot above 0 to do this.