Add files via upload
Found an example and went off it to make the two arraylists of objects that both control the visual representation and house the values I need. Right not Im working with the logic to make it so that a seatile displays the correct information depending on if it has a ship, if its been fired upon, and which board it is on. I also want to make it so that instead of a boolean denoting if a ship is placed, there is an int value to show that. that way I can distinguish between different ships (no ship==0, carrier=1, patrol boat=5, etc...) and check if a ship is sunk. From this I can make it so that the AI knows when to stop "hunting" around an area where it has hit like a real player would upon learning they sunk a ship.