Add files via upload
Added coordinate values to Seatiles so that I could hardocde in which computer tiles would have ships in them. Started work on a checker that sees if all seatiles of a certain ship value have been sunk. Tried a counter that tracks number of seatiles of each ship still present but may not be best approach. Weird to work with since I put everything within my Seatile object.